Moved dicom_times_to_ms() to utils
